Let lx_install eat gzipped ZFS send streams

+++ usr/src/lib/brand/lx/zone/lx_install.ksh
@@ -23,7 +23,7 @@ ZFS_SEED=""
 bad_usage() {
 	echo "LX zone install bad option"
 	echo "Available options are:"
-	echo "	-s <absolute-pathname>	Path to ZFS send stream"
+	echo "	-s <absolute-pathname>	Path to ZFS send stream or gzip thereof"
 	exit $ZONE_SUBPROC_USAGE
 }
 
@@ -46,7 +46,34 @@ fi
 # Set the ZONEPATH_DS variable so we know the zone's dataset.
 get_zonepath_ds $ZONEPATH
 
-zfs recv -F $ZONEPATH_DS < $ZFS_SEED
+# Do something based on whatever ZFS_SEED is.
+
+if [[ ! -f $ZFS_SEED ]]; then
+    echo "Seed file $ZFS_SEED not found."
+    # XXX KEBE SAYS maybe we can eat a snapshot name here, or even a
+    # Joyent-style UUID for direct snagging from Joyent's image
+    # servers.
+    bad_usage
+fi
+
+type=`file -b $ZFS_SEED | awk '{print $1}'`
+
+# For now, we are dependent on the output of file(1).
+# I'm being cheesy in checking the first word of file(1)'s output.
+if [[ $type == "ZFS" ]]; then
+    zfs recv -F $ZONEPATH_DS < $ZFS_SEED
+elif [[ $type == "gzip" ]]; then
+    gunzip -c $ZFS_SEED | zfs recv -F $ZONEPATH_DS
+else
+    echo "Seed file $ZFS_SEED not a ZFS receive (or compressed) one."
+    bad_usage
+fi
+
+if [[ $? != 0 ]]; then
+   echo "ZFS receive command failed ($?)."
+   exit $ZONE_SUBPROC_FATAL
+fi
+
 # One Joyent-ism we need to clean up.
 rmdir $ZONEPATH/cores
